Place "common" functions into cmd.go
